requirements.txt
# Core LLM and LangChain
langchain>=0.1.0
langchain-openai>=0.0.0
langchain-community>=0.0.0

# OpenAI
openai>=1.0.0

# Vector Store and Embeddings
chromadb>=0.4.0
sentence-transformers>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.0.0
